+++ app/public/components/sidebar/sidebarItem.js
@@ -19,7 +19,7 @@ export default function sidebarItem(model, pageName, label) {
     return h('.flex-wrap', [
         h('.page-title', {
             class: model.router.params.page === pageName ? 'selected' : '',
-            onclick: () => model.goToDefaultPageUrl(pageName),
+            onclick: () => model.navigation.goToDefaultPageUrl(pageName),
         },
         model.router.params.page === pageName
             ? h('div',
